+++ src/test/kotlin/com/github/hechtcarmel/jetbrainsindexmcpplugin/tools/navigation/GetSignatureToolUnitTest.kt
@@ -1,0 +1,12 @@
+package com.github.hechtcarmel.jetbrainsindexmcpplugin.tools.navigation
+
+import com.github.hechtcarmel.jetbrainsindexmcpplugin.constants.ToolNames
+import junit.framework.TestCase
+
+class GetSignatureToolUnitTest : TestCase() {
+    fun testNameAndSchema() {
+        val tool = GetSignatureTool()
+        assertEquals(ToolNames.GET_SIGNATURE, tool.name)
+        assertTrue(tool.inputSchema.containsKey("properties"))
+    }
+}